Refactoring: Move IFieldDefinition into new module metaschema.py.
