+++ opengever/propertysheets/api/post.py
@@ -2,44 +2,17 @@ from opengever.api.add import get_schema_validation_errors
 from opengever.propertysheets.definition import isidentifier
 from opengever.propertysheets.definition import PropertySheetSchemaDefinition
 from opengever.propertysheets.exceptions import InvalidSchemaAssignment
+from opengever.propertysheets.metaschema import IFieldDefinition
 from opengever.propertysheets.storage import PropertySheetSchemaStorage
 from plone import api
 from plone.protect.interfaces import IDisableCSRFProtection
 from plone.restapi.deserializer import json_body
 from plone.restapi.services import Service
-from plone.supermodel import model
 from zExceptions import BadRequest
 from zExceptions import Unauthorized
-from zope import schema
 from zope.interface import alsoProvides
 from zope.interface import implementer
-from zope.interface import provider
 from zope.publisher.interfaces import IPublishTraverse
-from zope.schema.interfaces import IContextSourceBinder
-from zope.schema.vocabulary import SimpleVocabulary
-
-
-@provider(IContextSourceBinder)
-def make_field_types_vocabulary(context):
-    return SimpleVocabulary.fromValues(
-        PropertySheetSchemaDefinition.FACTORIES.keys()
-    )
-
-
-class IFieldDefinition(model.Schema):
-
-    name = schema.TextLine(required=True)
-    field_type = schema.Choice(
-        required=True, source=make_field_types_vocabulary
-    )
-    title = schema.TextLine(required=False)
-    description = schema.TextLine(required=False)
-    required = schema.Bool(required=False)
-    values = schema.List(
-        default=None,
-        value_type=schema.TextLine(),
-        required=False,
-    )
 
 
 @implementer(IPublishTraverse)
